migrate jsonl utils
