+++ libs/firebase-utils/src/lib/firestore/index.ts
@@ -2,3 +2,4 @@ export * from './clear'
 export * from './import'
 export * from './export'
 export * from './emulator'
+export * from './jsonlArchive'
